pyproject.toml
[project]
name = "vici"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.1",
    "greenlet>=3.3.2",
    "inngest>=0.5.17",
    "opentelemetry-api>=1.40.0",
    "opentelemetry-exporter-otlp>=1.40.0",
    "opentelemetry-instrumentation-fastapi>=0.61b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.61b0",
    "opentelemetry-sdk>=1.40.0",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "sqlmodel>=0.0.37",
    "structlog>=25.5.0",
    "twilio>=9.10.2",
    "uvicorn[standard]>=0.41.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.5",
]
